requirements.txt
streamlit
langchain
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
langchain-community